requirements.txt
perlin-noise==1.12
pillow==10.3.0
pygame==2.5.2
